add bert data, refactor routes out of llm which dont use llm in path, add ability to get ungrouped demographic groups
